[move-package] MovePackage accepts CompiledModules by reference
Preparation for package upgrades, where to verify `CompiledModules`,
we need linkage and type origin tables, which are created as part of a
`MovePackage`, but currently, the `MovePackage` is created after
verification.

This change allows us to deserialize bytes to create
`CompiledModule`s, then build a `MovePackage` from those, and use that
to verify the deserialized modules.

Test Plan:

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```