[upgrades] adapter sets link context and table (#9932)
## Description

### Create MovePackage before verification

Preparation for package upgrades, where to verify `CompiledModules`,
we need linkage and type origin tables, which are created as part of a
`MovePackage`, but currently, the `MovePackage` is created after
verification.

This change allows us to deserialize bytes to create
`CompiledModule`s, then build a `MovePackage` from those by reference,
and use that to publish and verify the deserialized modules.

### Removes version parameter from MovePackage::new_initial

Get rid of `publish_and_verify_new_modules`,
`publish_and_verify_upgraded_modules` and `generate_package_id` in
favour of inlining them into their respective call-sites.  This is to
prepare for generating a `MovePackage` before calling
`publish_and_verify_modules` which requires having access to the list
of `CompiledModule`s after deserialization and address substitution.

In preparation for verification requiring a context that needs to be
derived from a `MovePackage`.

### LinkageView

Remove implementation of LinkageResolver for `StorageView`, and
introduce `LinkageView` as a wrapping type that holds linkage
information and delegates all other resolution queries to
`StorageView`.

Relies on move-language/move#1009 which adds support to `Session` for
resolvers that are richer than just a reference.

### Set linkage table for ExecutionContext

Add an API to set the linkage table based on a `MovePackage`, and
read from it when responding to `relocate` and `link_context` queries.

Setting the appropriate linkage context for calls into the VM and type
loading.  After setting the package's link context for a move call, we
also update its package ID to its runtime ID.

Sui Move bump, bringing in the following changes:

- Expose more of the loader to the adapter layer.
- More flexible compatibility checks
- ...and other fixes

### Backwards Compatibility

- Make sure the order of errors produced by publish remain consistent
when upgrades are not enabled as a feature, and don't perform any relinking.

### Hardening

Make it an error to try and overwrite an existing link context.  This
also uncovered a bug where we set the link context as part of taking
user events at the end of a move call, which could cause issues in the
case of:

- a publish
- ...with multiple modules
- ...multiple of which have initializers
- ...which emit events
- ...which come from a different package

## Test Plan

Existing tests and newly enabled upgrade tests:

```
$ cargo simtest
$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```

Fixes to the interleaved package upgrade test, resolving two issues:

1. The helper function for building the package, computing the digest,
   and the like was returning the wrong IDs (it was returning original
   package IDs and not the new dependency IDs).
2. We were trying to use the new object ID for the dependency as both
   the `published-at` and address value for the package.

New enabled tests:

- Calling into an upgraded package.
- Referencing types introduced during upgrades.